+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added
   - Add support for `this.state` initialization in constructor
   - Add support for `defaultProps`
+  - Add support for `const { ... } = this.state`
 - Misc
   - Heavily refactored internal analysis